examples/python/requirements.txt
grpcio==1.71.1
grpcio-tools==1.71.1
protobuf==5.29.2